+++ dragonchain/exceptions.py
@@ -57,10 +57,6 @@ class OpenFaasException(DragonchainException):
     """Exception raised when OpenFaaS returns with error status"""
 
 
-class LabChainForbiddenException(DragonchainException):
-    """Exception raised when lab chain action is not allowed"""
-
-
 class NotFound(DragonchainException):
     """Exception raised when object is not found"""
 
@@ -153,6 +149,10 @@ class StorageError(DragonchainException):
 
 class SanityCheckFailure(DragonchainException):
     """Exception raised when sanity check fails"""
+
+
+class InterchainPublishError(DragonchainException):
+    """Exception raised when an interchain publish action fails"""
 
 
 class InterchainConnectionError(DragonchainException):