+++ dragonchain/lib/dto/eth.py
@@ -234,7 +234,17 @@ class EthereumNetwork(model.InterchainModel):
         """
         return base64.b64encode(self.priv_key.to_bytes()).decode("ascii")
 
-    def _publish_transaction(self, transaction_payload: str) -> str:
+    def publish_transaction(self, signed_transaction: str) -> str:
+        """Publish an already signed transaction to this network
+        Args:
+            signed_transaction: The already signed transaction from self.sign_transaction
+        Returns:
+            The hex string of the published transaction hash
+        """
+        _log.debug(f"[ETH] Publishing transaction {signed_transaction}")
+        return self.w3.toHex(self.w3.eth.sendRawTransaction(signed_transaction))
+
+    def _publish_l5_transaction(self, transaction_payload: str) -> str:
         """Publish a transaction to this network with a certain data payload
         Args:
             transaction_payload: The arbitrary data to send with this transaction
@@ -251,7 +261,7 @@ class EthereumNetwork(model.InterchainModel):
             }
         )
         # Send signed transaction
-        return self.w3.toHex(self.w3.eth.sendRawTransaction(signed_transaction))
+        return self.publish_transaction(signed_transaction)
 
     def _calculate_transaction_fee(self) -> int:
         """Get the current gas price estimate