+++ dragonchain/webserver/helpers.py
@@ -16,9 +16,7 @@
 # language governing permissions and limitations under the Apache License.
 
 import json
-import os
-import functools
-from typing import Tuple, Dict, Any, Callable, Iterable, TYPE_CHECKING
+from typing import Tuple, Dict, Any, Iterable, TYPE_CHECKING
 
 import fastjsonschema
 from werkzeug import exceptions as werkzeug_exceptions
@@ -32,18 +30,6 @@ if TYPE_CHECKING:
     from dragonchain.lib.types import custom_index  # noqa: F401
 
 _log = logger.get_logger()
-_is_lab = os.environ.get("LAB") == "true"
-
-
-def DisabledForLab(route: Callable) -> Callable:  # noqa: N802
-    @functools.wraps(route)
-    def decorator(*args: Any, **kwargs: Any) -> Any:
-        if _is_lab:
-            raise exceptions.LabChainForbiddenException
-        else:
-            return route(*args, **kwargs)
-
-    return decorator
 
 
 def flask_http_response(status: int, data: Any) -> Tuple[str, int, Dict[str, str]]:
@@ -88,10 +74,6 @@ NOT_ACCEPTING_VERIFICATIONS = format_error("NOT_ACCEPTING_VERIFICATIONS", "Not c
 INVALID_TRANSACTION_TYPE = format_error(
     "INVALID_TRANSACTION_TYPE", "The transaction type you are attempting to use either does not exist or is invalid."
 )
-ACTION_FORBIDDEN_LAB_CHAIN = format_error(
-    "ACTION_FORBIDDEN_LAB_CHAIN",
-    "This feature is disabled for Labs. If you are interested in this feature, please visit https://dragonchain.com/pricing",
-)
 INTERNAL_SERVER_ERROR = format_error("INTERNAL_SERVER_ERROR", "Unexpected error occurred")
 
 
@@ -129,6 +111,11 @@ def validation_exception(exception: Exception) -> dict:
 def bad_request(exception: Exception) -> dict:
     message = error_reporter.get_exception_message(exception)
     return format_error("BAD_REQUEST", message)
+
+
+def interchain_publish_error(exception: Exception) -> dict:
+    message = error_reporter.get_exception_message(exception)
+    return format_error("INTERCHAIN_PUBLISH_ERROR", message)
 
 
 def webserver_error_handler(exception: Exception) -> Tuple[str, int, Dict[str, str]]:  # noqa C901
@@ -180,9 +167,6 @@ def webserver_error_handler(exception: Exception) -> Tuple[str, int, Dict[str, s
     elif isinstance(exception, exceptions.BadDockerAuth):
         status_code = 400
         surface_error = BAD_DOCKER_AUTH_ERROR
-    elif isinstance(exception, exceptions.LabChainForbiddenException):
-        status_code = 403
-        surface_error = ACTION_FORBIDDEN_LAB_CHAIN
     elif isinstance(exception, werkzeug_exceptions.MethodNotAllowed):
         status_code = 405
         surface_error = METHOD_NOT_ALLOWED
@@ -192,6 +176,9 @@ def webserver_error_handler(exception: Exception) -> Tuple[str, int, Dict[str, s
     elif isinstance(exception, exceptions.OpenFaasException):
         status_code = 500
         surface_error = OPENFAAS_ERROR
+    elif isinstance(exception, exceptions.InterchainPublishError):
+        status_code = 500
+        surface_error = interchain_publish_error(exception)
     else:
         status_code = 500
         surface_error = INTERNAL_SERVER_ERROR