+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 4.4.0
+
+- **Feature:**
+  - Modify L4 block schema to include chain name in the L4 block header
+  - Add new endpoint `DELETE /v1/contract/txn_type/<txn_type>` for deleting smart contracts by transaction type
+  - Add new endpoint `POST /v1/interchains/transaction/publish` for publishing a signed interchain transaction
+- **Bugs:**
+  - Fix issue where L5 recovery queue would not process when trying to resolve already resolved claim-checks
+- **Packaging:**
+  - Update redis, web3, and boto3 dependencies
+  - Update redisearch in helm chart to 1.6.7
+  - Update fwatchdog to 0.18.10 for OpenFaaS smart contracts
+- **Development:**
+  - Removed all references about lab chain
+  - Remove codeowners
+
 ## 4.3.3
 
 - **Bugs:**